Created Cell object
